Resulation&UI Update

+++ Chemistry/src/com/chemistrylab/layer/Range.java
@@ -1,8 +1,26 @@
 package com.chemistrylab.layer;
 
+/**
+ * <p>
+ * A range, which an area meaning a layer/component position.
+ * </p>
+ * range.x0 range.y0<br/>
+ * *--------+<br/>
+ * |&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
+ * &nbsp;|<br/>
+ * |&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
+ * &nbsp;|<br/>
+ * |&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
+ * &nbsp;|<br/>
+ * +--------*<br/>
+ * range.x1 range.y1
+ * 
+ * @author Nickid2018
+ *
+ */
 public class Range {
-	public int x0;
-	public int y0;
-	public int x1;
-	public int y1;
+	public float x0;
+	public float y0;
+	public float x1;
+	public float y1;
 }
