+++ Chemistry/src/com/chemistrylab/layer/component/Slidable.java
@@ -3,9 +3,13 @@ package com.chemistrylab.layer.component;
 public interface Slidable {
 
 	void debugRender();
+
 	void render();
-	void setNowPositon(int x0,int x1,int y0,int y1);
+
+	void setNowPositon(float x0, float x1, float y0, float y1);
+
 	void onMouseEvent();
+
 	boolean checkRange(int x, int y);
 
 }
